Name the peripherals after their classes and hold every header to the budget
The three parallel-LED backends are named for the classes they contain, so a file and its type
agree. Every header behind the drivers page now states its contract inside the documented comment
budget, which the check enforces absolutely: the grandfather list is gone.

perf: desktop fps 8000 → 7874 (-126) · tick_us 125 → 127 (+2) · peripheral_grid_sweep p95 286 → 291 (+5) · loc.light 36453 → 32887 (-3566) · comments.light 13870 → 10299 (-3571)

**Core**
- Control.cpp walks a Select's options in the type of the bound it compares against. A uint8_t
  counter against a wider signed max is a wrap that only today's option count keeps unreachable,
  and the suppression that hid it outlived its reason.

**Light domain**
- MultiPinLedDriver.h, MoonLedDriver.h and ParlioLedDriver.h become I80Peripheral.h,
  MoonI80Peripheral.h and ParlioPeripheral.h. Each held one LedPeripheral backend, never a driver:
  the one ParallelLedDriver drives all three. Their tests and the persistence comment follow.
- Fifteen driver headers swept to the budget: 6,665 lines of comment removed, no code changed.
  A constraint a maintainer needs stays as one line at the code; what a user decides with moved to
  the drivers page; how a rule came to be broken is in git, so it is deleted. Every public member
  carries a one-line ///.

**Platform**
- The desktop config write's owner-only open is hoisted out of the atomic-write path and used by
  the mount probe too. Plain fopen takes the process umask, so on a permissive one the probe was
  world-writable while it existed. It holds nothing, but a file this code creates should not be
  the loosest thing in the directory.

**Scripts/MoonDeck**
- docgen_baseline.txt is deleted, with --baseline and the tolerate-at-recorded-size logic. A
  baseline hid findings at the size they happened to be, which is the opposite of a limit.
- The one-line cap and the 20-word cap now cover `//` as well as ///. Re-spelling a fifty-line
  member comment as `//` satisfied every other rule and left the file the same length, which is
  what a first pass through these headers produced. A file-level block above the first class is
  exempt, being the non-Doxygen sibling of the class comment.
- gen_api.py reports a generated page nothing links to, so a header that gains a page without a
  reference is visible rather than orphaned.

**Docs/CI**
- documentation-standards.md states the budget as one table: 10 lines of class ///, a 20-line
  @moreinfo appendix, one line for any other run, 20 words on a comment line, one line for a `//`
  run beside code, and a /// on every public member.
- The drivers page gains the HUB75 pin maps, the panel-scanning explanation, the '595 expander
  limits and the loopback wiring, each moved out of a header appendix.
- backlog-core records that check_nonblocking's baseline has inverted: 107 entries against 240
  live sites, so files it never covered report every site as new. Removing it is cheap; making
  the check a gate waits for the hot-path work above it.

+++ src/core/Control.cpp
@@ -202,10 +202,11 @@ void writeControlMetadata(JsonSink& sink, const ControlDescriptor& c) {
         case ControlType::Select: {
             sink.append(",\"options\":[");
             auto* options = reinterpret_cast<const char* const*>(c.aux);
-            // addSelect takes a uint8_t option count, so c.max can never exceed 255 and the
-            // counter cannot wrap.
-            // NOLINTNEXTLINE(bugprone-too-small-loop-variable)
-            for (uint8_t o = 0; o < c.max; o++) {
+            // int32_t, the type of the bound it is compared against: a uint8_t counter against a
+            // wider signed max is a wrap that only today's uint8_t option count keeps unreachable,
+            // and the suppression that hid it outlived the reason for it. Same form as the hash
+            // walk in MoonModule.h.
+            for (int32_t o = 0; o < c.max; o++) {
                 if (o > 0) sink.append(",");
                 // Escaped, not a raw %s: most option lists are our own literals, but the panel-card
                 // interface Select carries OS-supplied adapter descriptions, and one containing a
